ExportController using DataExporter

+++ src/Libraries/SmartStore.Services/DataExchange/IDataExporter.cs
@@ -13,12 +13,9 @@ namespace SmartStore.Services.DataExchange
 	{
 		DataExportResult Export(DataExportRequest request, CancellationToken cancellationToken);
 
-		// Handle model conversion for grid in backend's controller
 		IList<dynamic> Preview(DataExportRequest request);
 
-		// useful for decision making whether export should
-		// be processed sync or async
-		long GetDataCount(DataExportRequest request);
+		int GetDataCount(DataExportRequest request);
 	}
